pyproject.toml
[tool.poetry]
name = "pyexoyone"
version = "0.0.0"
description = "Python library to configure and control the Exoy(tm) ONE."
authors = ["Avi Miller <me@dje.li>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/Djelibeybi/pyExoyOne"
documentation = "https://github.com/Djelibeybi/pyExoyOne"
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries",
]
packages = [
    { include = "exoyone", from = "src" },
    { include = "moxyone", from = "tests"}
]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/Djelibeybi/pyexoyone/issues"
"Changelog" = "https://github.com/Djelibeybi/pyexoyone/blob/main/CHANGELOG.md"

[tool.poetry.scripts]
exoyone = "exoyone.__main__:app"
moxyone = "moxyone.__main__:app"

[tool.poetry.dependencies]
python = "^3.12"
rich = ">=10"
zeroconf = "^0.132.2"
camel-converter = "^3.1.2"
typer = "^0.12.3"
asyncio-dgram = "^2.2.0"
backoff = "^2.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
ruff = "^0.5.6"
ruff-lsp = "^0.0.54"
pre-commit = "^3.8.0"
pytest-asyncio = "^0.23.8"
pytest-xprocess = "^1.0.2"
pytest-cov = "^5.0.0"
mypy = "^1.11.1"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.6.0"
mkdocs-material = "^9.5.31"
mkdocs-typer = "^0.0.3"
mkdocstrings = {extras = ["python"], version = "^0.25.2"}

[tool.semantic_release]
version_toml = ["pyproject.toml:tool.poetry.version"]
version_variables = [
    "src/exoyone/__init__.py:__version__",
]
build_command = "pip install poetry && poetry build"

[tool.semantic_release.changelog]
exclude_commit_patterns = [
    "chore*",
    "ci*",
]

[tool.semantic_release.changelog.environment]
keep_trailing_newline = true

[tool.semantic_release.branches.main]
match = "main"

[tool.semantic_release.branches.noop]
match = "(?!main$)"
prerelease = true

[tool.pytest.ini_options]
addopts = """\
    -v
    -Wdefault
    --capture=tee-sys
    --cov=exoyone
    --cov-report=term-missing:skip-covered
    --cov-report=xml
    --cov-report=html
    --cov-branch
    """
pythonpath = ["src"]
log_cli = true
filterwarnings = [
    "error",
    "ignore::UserWarning",
    # note the use of single quote below to denote "raw" strings in TOML
    'ignore:function ham\(\) is deprecated:DeprecationWarning',
]


[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "@overload",
    "if TYPE_CHECKING",
    "raise NotImplementedError",
    'if __name__ == "__main__":',
]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
ignore = [
    "D203", # 1 blank line required before class docstring
    "D212", # Multi-line docstring summary should start at the first line
    "D100", # Missing docstring in public module
    "D104", # Missing docstring in public package
    "D107", # Missing docstring in `__init__`
    "D401", # First line of docstring should be in imperative mood
    "S311", # Standard pseudo-random generators are not suitable for cryptographic purposes
]
select = [
    "B",   # flake8-bugbear
    "D",   # flake8-docstrings
    "C4",  # flake8-comprehensions
    "S",   # flake8-bandit
    "F",   # pyflake
    "E",   # pycodestyle
    "W",   # pycodestyle
    "UP",  # pyupgrade
    "I",   # isort
    "RUF", # ruff specific
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*" = [
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "S101",
]
"setup.py" = ["D100"]
"conftest.py" = ["D100"]
"docs/conf.py" = ["D100"]

[tool.ruff.lint.isort]
known-first-party = ["exoyone", "tests"]

[tool.mypy]
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
mypy_path = "src/"
no_implicit_optional = true
show_error_codes = true
warn_unreachable = true
warn_unused_ignores = true
exclude = [
    'docs/.*',
    'setup.py',
]

[[tool.mypy.overrides]]
module = "tests.*"
allow_untyped_defs = true

[[tool.mypy.overrides]]
module = "docs.*"
ignore_errors = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
